agrego control de luces por switch (exterior, internas, noche))

+++ Truma_EBL_Bios_V100/Truma_EBL_Bios_V100/Core/Inc/EmptyingValve.h
@@ -17,7 +17,7 @@
 class EmptyingValve : public Devices, public SystemDevices
 {
 private:
-    unsigned long long expirationTime;
+    uint64_t expirationTime;
     int state;
 		bool stateChange;
 public:
