+++ Truma_EBL_Bios_V100/Truma_EBL_Bios_V100/Core/Inc/FreshWater_1.h
@@ -17,8 +17,8 @@
 class FreshWater1 : public Devices, public SystemDevices
 {
 private:
-    unsigned long long expirationTime;
-		unsigned long long powerTime;
+    uint64_t expirationTime;
+		uint64_t powerTime;
     int state;
 		float levelFW;
 		int counterFW;
