+++ Truma_EBL_Bios_V100/Truma_EBL_Bios_V100/Core/Inc/Fridge.h
@@ -17,7 +17,7 @@
 class Fridge : public Devices, public SystemDevices
 {
 private:
-    unsigned long long expirationTime;
+    uint64_t expirationTime;
     int state;
 		bool stateChange;
 public:
