+++ Truma_EBL_Bios_V100/Truma_EBL_Bios_V100/Core/Inc/LeisureBattery1.h
@@ -18,7 +18,7 @@
 class LeisureBattery1 : public Devices, public SystemDevices, public LinClients 
 {
 private:
-    unsigned long long expirationTime;
+    uint64_t expirationTime;
 public:
     LeisureBattery1(short _idTopic);
